+++ packages/frontend/src/Theme.ts
@@ -1,20 +1,20 @@
-import {type DefaultTheme} from 'styled-components';
+import { type DefaultTheme } from 'styled-components';
 
 const Theme: DefaultTheme = {
-  borderColor: '#bbb',
-  color: 'hsl(0, 0%, 7%)',
-  fontFamilies: {
-    RobotoFlex: 'Roboto Flex, sans-serif;',
-    RobotoMono: 'Roboto Mono, monospace;',
-  },
-  fontSize: '16px',
-  printSize: '10pt',
-  fontWeight: '400',
-  tfBlue: 'hsl(210, 61%, 22%)',
-  tfBlueHighlight: 'hsl(210, 58%, 90%)',
-  tfYellowCTA: 'hsl(47, 100%, 62%)',
-  tfYellowCTAhover: 'hsl(47, 100%, 50%)',
-  white: 'hsl(0, 0%, 100%)',
+	borderColor: '#bbb',
+	color: 'hsl(0, 0%, 7%)',
+	fontFamilies: {
+		RobotoFlex: 'Roboto Flex, sans-serif;',
+		RobotoMono: 'Roboto Mono, monospace;',
+	},
+	fontSize: '16px',
+	printSize: '10pt',
+	fontWeight: '400',
+	tfBlue: 'hsl(210, 61%, 22%)',
+	tfBlueHighlight: 'hsl(210, 58%, 90%)',
+	tfYellowCTA: 'hsl(47, 100%, 62%)',
+	tfYellowCTAhover: 'hsl(47, 100%, 50%)',
+	white: 'hsl(0, 0%, 100%)',
 };
 
 export default Theme;
